layers remake

+++ src/routes/bro-routers.ts
@@ -1,10 +1,4 @@
 import express, { Response } from 'express'
-import { RequestWithBody, RequestWithParams, RequestWithParamsAndBody, RequestWithQuery } from '../types'
-import { CreateBroModel } from '../models/CreateBroModel'
-import { GetBroModel } from '../models/GetBroQueryModel'
-import { UpdateBroModel } from '../models/UpdateBroModel'
-import { ViewBroModel } from '../models/ViewBroModel'
-import { URIparamId } from '../models/URIParamIdModel'
 import { HTTP_STATUSES } from '../utils'
 import { body } from 'express-validator'
 import { inputValidationMiddleWare } from '../middlewares/input-validation-mw'
@@ -15,50 +9,51 @@ import { brothersServise } from '../domain/brothers-servise'
 export const broRouters = () => {
 
     const router = express.Router()
-    const titleValidation = body('title').trim().isLength({ min: 2, max: 10 }).withMessage("title length should be from two to then symbols")
+    const titleValidation = body('userName').trim().isLength({ min: 2, max: 10 }).withMessage("userName length should be from two to then symbols")
 
-    router.post('/', titleValidation, inputValidationMiddleWare, async (req: RequestWithBody<CreateBroModel>, res: any) => {
-        return res.status(HTTP_STATUSES.CREATED_201).send(await brothersServise.createBrother(req.body.title))
-    })
+    router.post('/', titleValidation, inputValidationMiddleWare,
+        async (req: Request<{}, {}, { userName: string, bio: string }>, res: Response) => {
+            const newUser = await brothersServise.createBrother(req.body.title)
+            return res.status(HTTP_STATUSES.CREATED_201).send(newUser)
+        })
 
+    router.put('/:id', titleValidation, inputValidationMiddleWare,
+        async (req: Request<{ id: string }, { userName: string, bio: string }>, res: Response) => {
+            const isUpdated = await brothersServise.updateBro(+req.params.id, req.body.bio, req.body.userName) // check the order updateBre params
 
-    router.get('/', async (req: RequestWithQuery<GetBroModel>, res: Response<ViewBroModel[]>) => {
-        const findedBro = await brothersServise.findBro(req.query.title?.toString())
-        res.send(findedBro)
-    })
+            if (isUpdated) {
+                res.sendStatus(HTTP_STATUSES.NO_CONTENT_204)
+            } else {
+                res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
+            }
+        })
 
-    router.get('/:id', async (req: RequestWithParams<URIparamId>, res: Response) => {
-        const findedBro = await brothersServise.getBroById(+req.params.id)
+    router.get('/',
+        async (req: Request, res: Response) => {
+            const findedBro = await brothersServise.findBro(req.query.title?.toString())
+            res.send(findedBro)
+        })
 
-        if (findedBro) {
-            res.json({ id: findedBro.id, title: findedBro.title })
-        } else {
-            res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
-            return
-        }
-    })
+    router.get('/:id',
+        async (req: Request<{ id: string }, { userName: string, bio: string }>, res: Response) => {
+            const findedBro = await brothersServise.getBroById(+req.params.id)
+            if (findedBro) {
+                res.json(findedBro)
+            } else {
+                res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
+                return
+            }
+        })
 
-
-
-    router.put('/:id', titleValidation, inputValidationMiddleWare, async (req: RequestWithParamsAndBody<URIparamId, UpdateBroModel>, res) => {
-        const isUpdated = await brothersServise.updateBrotherName(+req.params.id, req.body.title)
-
-        if (isUpdated) {
-            const bro = brothersServise.getBroById(+req.params.id)
-            res.sendStatus(HTTP_STATUSES.NO_CONTENT_204).send(bro)
-        } else {
-            res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
-        }
-    })
-
-    router.delete('/:id', async (req: RequestWithParams<URIparamId>, res) => {
-        const broForDelete = await brothersServise.deleteBro(+req.params.id)
-
-        if (!broForDelete) {
-            res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
-        } else {
-            res.sendStatus(HTTP_STATUSES.NO_CONTENT_204).send(brothersServise.getBroById(+req.params.id))
-        }
-    })
+    router.delete('/:id',
+        async (req: Request<{ id: string }>, res: Response) => {
+            const isBroExist = await brothersServise.getBroById(+req.params.id)
+            if (isBroExist) {
+                res.sendStatus(HTTP_STATUSES.NO_CONTENT_204).send(isBroExist)
+                brothersServise.deleteBro(+req.params.id)
+            } else {
+                res.sendStatus(HTTP_STATUSES.NOT_FOUND_404)
+            }
+        })
     return router
 }
